Move <ChevronRight /> to more central folder
